health decay with version v1 of non-linear updateHP
updateHP is modeled as a 1st order system step response starting at a given level.
The required amount of food to stay in a given level corresponds to the time constant "tau" of this step response. For staying at the same level, the hp should reach 63% of the total level "width", and 95% to go reach the level above.
For now, every day start at the bottom of a given level.

Critical state not correctly implemented yet