Team2 logging (#263)
* Reset dead agents instead of removing them

* fixed a bug in Run that cause multiple updates within one day

* Fix bug for which table updates when no action was performed

* Added function to export Q Table to CSV file

* Added exportQTable() function

Agent Q Table is exported as a CSV file called "qtable.csv". Currently runs at end of simulation. The day count in config.go must be changed if it is changed in config.json.

* Added exportPolicies().

Policies are exported to policies.csv

* Removed .Day() hack from baseAgent.go

* Reset dead agents instead of removing them

* fixed a bug in Run that cause multiple updates within one day

* Fix bug for which table updates when no action was performed

* Added function to export Q Table to CSV file

* Added exportQTable() function

Agent Q Table is exported as a CSV file called "qtable.csv". Currently runs at end of simulation. The day count in config.go must be changed if it is changed in config.json.

* Added exportPolicies().

Policies are exported to policies.csv

* Removed .Day() hack from baseAgent.go

* Removed reincarnation features

* Fixed policy test

* Fixed lint errors

Co-authored-by: laiwenq <wenqianglai@hotmail.com>